refs #9 add negative test cases on TimeTree::Client and do refactor.
