+++ test/testdata/401.json
@@ -1,0 +1,5 @@
+{
+  "type": "https://developers.timetreeapp.com/en/docs/api#authentication",
+  "title": "Unauthorized",
+  "status": 401
+}
