+++ test/testdata/404_calendar.json
@@ -1,0 +1,5 @@
+{
+  "type": "https://developers.timetreeapp.com/en/docs/api#get-calendarscalendar_id",
+  "title": "Not Found",
+  "status": 404
+}
